one inheritance error, fixing xcode project

+++ FaceSubstitution/src/FaceBlinkRecorder.cpp
@@ -121,7 +121,7 @@ VideoFrame * FaceBlinkRecorder::getNextVideoFrame(){
     return vframe;
 }
 
-int FaceBlinkRecorder::getFps(){
+float FaceBlinkRecorder::getFps(){
 	return fps;
 }
